lttng-ctl: move lttng_event functions to a new file
