+++ src/lib/lttng-ctl/Makefile.am
@@ -6,7 +6,7 @@ lib_LTLIBRARIES = liblttng-ctl.la
 
 liblttng_ctl_la_SOURCES = lttng-ctl.c snapshot.c lttng-ctl-helper.h \
 		lttng-ctl-health.c save.c load.c deprecated-symbols.c \
-		channel.c rotate.c
+		channel.c rotate.c event.c
 
 liblttng_ctl_la_LDFLAGS = \
 		$(LT_NO_UNDEFINED)
